* Properly cast some things in sipXtackLib that is obvious to tell what type is intended. * when using UtlBoolean, use TRUE and FALSE instead of true and false.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9756 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXtackLib/src/net/SipPresenceEvent.cpp
@@ -214,7 +214,7 @@ void SipPresenceEvent::parseBody(const char* bodyBytes)
                      priority = ucElement->Attribute("priority");
                   }
                   
-               pTuple->setContact(contact, atof(priority));
+               pTuple->setContact(contact, (float)atof(priority));
                }
              
                // Insert it into the list
